renderer: ensure makecurrent isnt called twice
depending on the attachment and commitstate we might end up calling
makecurrent twice on an already current context, and on destruction we
also make it current yet again even tho it was current. drivers doesnt
keep track of this and is very inefficient when calling this on itself
all the time. add a bool and only call it when it has changed or require
change.